refactor: update pages imports, exports
