+++ src/pages/Department.js
@@ -2,16 +2,18 @@ import { useEffect, useState } from "react";
 import { useAuthHeader } from "react-auth-kit";
 import { Dialog } from "@mui/material";
 
+import { handleGetDepartmentsList, handleGetUsersList } from "services";
 import { Breadcrumb, PageContainer } from "components/Common";
-import {
-  ConstructionBreadcrumb,
-  ConstructionCreate,
-  ConstructionList,
-} from "components/Construction";
 import { Header } from "components/Header";
-import { handleGetConstructionList, handleGetUsersList } from "services";
+import {
+  DepartmentBreadcrumb,
+  DepartmentCreate,
+  DepartmentEdit,
+  DepartmentList,
+  DepartmentView,
+} from "components/Department";
 
-const ConstructionsPage = () => {
+const Department = () => {
   const authHeader = useAuthHeader();
 
   useEffect(() => {
@@ -19,26 +21,26 @@ const ConstructionsPage = () => {
       header: { Authorization: authHeader() },
       setState: setState,
     });
-    handleGetConstructionList({
+    handleGetDepartmentsList({
       header: { Authorization: authHeader() },
       setState: setState,
-    })
+    });
   }, []);
 
   const [state, setState] = useState({
     view: "",
-    selectedConstruction: null,
+    selectedDepartment: null,
     openDialog: false,
     users: [],
-    constructions: [],
+    departments: [],
   });
 
-  const openDialog = (view, construction = null) => {
+  const openDialog = (view, department = null) => {
     setState((prev) => ({
       ...prev,
       view,
       openDialog: true,
-      selectedConstruction: construction,
+      selectedDepartment: department,
     }));
   };
 
@@ -47,24 +49,38 @@ const ConstructionsPage = () => {
   };
 
   const views = {
-    list: <ConstructionList openDialog={openDialog} constructions={state.constructions} />,
+    list: (
+      <DepartmentList departments={state.departments} openDialog={openDialog} />
+    ),
     create: (
-      <ConstructionCreate
+      <DepartmentCreate
         onClose={closeDialog}
         users={state.users}
         setState={setState}
       />
     ),
-    update: null,
-    view: null,
+    update: (
+      <DepartmentEdit
+        department={state.selectedDepartment}
+        users={state.users}
+        setState={setState}
+        onClose={closeDialog}
+      />
+    ),
+    view: (
+      <DepartmentView
+        department={state.selectedDepartment}
+        onClose={closeDialog}
+      />
+    ),
   };
 
   return (
     <>
       <Header />
       <PageContainer>
-        <Breadcrumb title="Obras">
-          <ConstructionBreadcrumb />
+        <Breadcrumb title="Departamentos">
+          <DepartmentBreadcrumb />
         </Breadcrumb>
         {views.list}
       </PageContainer>
@@ -79,4 +95,4 @@ const ConstructionsPage = () => {
   );
 };
 
-export default ConstructionsPage;
+export { Department };
