+++ src/pages/ForgotPassword.js
@@ -116,4 +116,4 @@ const ForgotPassword = () => {
   );
 };
 
-export default ForgotPassword;
+export { ForgotPassword };
