+++ src/pages/Home.js
@@ -160,4 +160,4 @@ const Home = () => {
   );
 };
 
-export default Home;
+export { Home };
