+++ src/pages/index.js
@@ -1,5 +1,7 @@
-export { default as Home } from "./Home";
-export { default as Login } from "./Login";
-export { default as User } from "./User";
-export { default as ForgotPassword } from "./ForgotPassword";
-export { default as Profile } from "./Profile";
+export * from "./Construction";
+export * from "./Department";
+export * from "./ForgotPassword";
+export * from "./Home";
+export * from "./Login";
+export * from "./Profile";
+export * from "./User";
